Added initial scripts
Two scripts, a parser function file and the main file has been added.

Parser function file parses any additional inputs to the main file.

The main file, export_pretty_fig(), is the main function to be
called. It makes some common adjustments to the figure, and then exports
the figure to a PDF via export_fig.

Additionally, it manipulates some default options for matlab2tikz so
that it can be configured by default to as close as what one gets from
export_fig.

As of this commit, defaults have been chosen so that re-sizing the tikz
figure is easy via use of the tikzscale package in LaTeX. This ensures
that the labels are automatically adjusted to the default text size of
the document, and allowing custom axis ratio when including the tikz
figure in LaTeX.